Refactor functions to combat module
